backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# PDF Processing
PyMuPDF==1.23.8
pdfplumber==0.10.3
pdf2image==1.16.3
Pillow==10.1.0

# OCR
pytesseract==0.3.10

# Table Extraction
openpyxl==3.1.2

# Vector Store & Embeddings
faiss-cpu==1.9.0
sentence-transformers==2.2.2

# LLM Integration
langchain==0.1.0
langchain-community==0.0.10
openai==1.6.1

# Database
sqlalchemy==2.0.23
aiosqlite==0.19.0

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
python-magic-bin==0.4.14
tqdm==4.66.1
numpy==1.26.2
pandas==2.1.4
jinja2==3.1.2

# Logging & Monitoring
loguru==0.7.2
psutil==5.9.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
